add BinaryShardedJedisBase class and Sharding interface
